myplatform-backend/requirements.txt
Django>=3.2,<4.0
djangorestframework
psycopg2-binary
gunicorn
django-cors-headers
drf-yasg
python-decouple
boto3
xlsxwriter
openpyxl
pandas
pytest
pytest-django
factory-boy
coverage
